+++ ref-impl/src/com-api/com-dll/CAAFInProcServer.cpp
@@ -1,6 +1,6 @@
 //=---------------------------------------------------------------------=
 //
-// $Id: CAAFInProcServer.cpp,v 1.25 2005/04/06 13:50:34 stuart_hc Exp $ $Name:  $
+// $Id: CAAFInProcServer.cpp,v 1.26 2006/06/15 19:53:19 tbingham Exp $ $Name:  $
 //
 // The contents of this file are subject to the AAF SDK Public
 // Source License Agreement (the "License"); You may not use this file
@@ -199,7 +199,7 @@ HRESULT CAAFInProcServer::Init
 
 
   // Compute the size of the table.
-  size_t i = 0;
+  ULONG i = 0;
   while (pObjectInfo[i].pCLSID)
     ++i;
   _objectCount = i;
@@ -390,7 +390,7 @@ HRESULT AAFGetLibraryInfo(HINSTANCE hInstance, char **pServerPath, char **pServe
 {
 	HRESULT rc = S_OK;
 	char path[MAX_PATH];
-	int pathLen = 0;
+	size_t pathLen = 0;
 	char *pDirSeparator = 0;
